+++ backend/app/db/models/order_fulfillment_metrics.py
@@ -1,5 +1,5 @@
 from app.db.models.base_class import Base
-from sqlalchemy import Column, Float, Integer
+from sqlalchemy import Column, Integer
 
 
 class OrderFulfillmentMetrics(Base):
@@ -8,7 +8,7 @@ class OrderFulfillmentMetrics(Base):
     order_id = Column(Integer, primary_key=True)
     item_id = Column(Integer, nullable=False)
     mart_id = Column(Integer, nullable=False)
-    fulfillment_time_minutes = Column(Float, nullable=False)
+    fulfillment_time_minutes = Column(Integer, nullable=False)
 
     def __repr__(self):
         return f"<OrderMetrics(order={self.order_id}, time={self.fulfillment_time_minutes})>"
